Implement get fire arms from data base by id and get all fire arms.

+++ src/main/webapp/model/Calibre.java
@@ -9,24 +9,19 @@ public class Calibre {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
 
-    private int id;
+    private long id;
 
     private double size;
-
-
-
 
 
     public Calibre() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-
-
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
