+++ src/main/webapp/model/Producer.java
@@ -10,15 +10,14 @@ public class Producer {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
 
-    private int id;
-
+    private long id;
 
     private String producerName;
 
     public Producer() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -26,7 +25,7 @@ public class Producer {
         return producerName;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
